requirements.txt
aiogram==3.3.0
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.1